+++ src/sveltekit/sveltekit-wizard.ts
@@ -140,10 +140,10 @@ without SvelteKit's builtin observability.`,
     }
   }
 
-  Sentry.setTag(
-    'svelte-version',
-    getSvelteVersionBucket(getPackageVersion('svelte', packageJson)),
+  const svelteVersionBucket = getSvelteVersionBucket(
+    getPackageVersion('svelte', packageJson),
   );
+  Sentry.setTag('svelte-version', svelteVersionBucket);
 
   const projectData = await getOrAskForProjectData(
     options,
@@ -212,12 +212,13 @@ without SvelteKit's builtin observability.`,
 
   if (shouldCreateExamplePage) {
     try {
-      await traceStep('create-example-page', () =>
+      await traceStep('create-example-page', async () =>
         createExamplePage(svelteConfig, {
           selfHosted,
           url: sentryUrl,
           orgSlug: selectedProject.organization.slug,
           projectId: selectedProject.id,
+          isUsingSvelte5: ['5.x', '>5.x'].includes(svelteVersionBucket),
         }),
       );
     } catch (e: unknown) {