+++ src/sveltekit/templates.ts
@@ -138,6 +138,7 @@ export function getSentryExampleSveltePage(options: {
   url: string;
   orgSlug: string;
   projectId: string;
+  isUsingSvelte5: boolean;
 }) {
   const issuesPageLink = options.selfHosted
     ? `${options.url}organizations/${options.orgSlug}/issues/?project=${options.projectId}`
@@ -152,9 +153,7 @@ Feel free to delete this file and the entire sentry route.
   import * as Sentry from '@sentry/sveltekit';
   import { onMount } from 'svelte';
   
-  // Svelte Runes (requires Svelte 5)
-  // let hasSentError = $state(false);
-  let hasSentError = false;
+  let hasSentError = ${options.isUsingSvelte5 ? '$state(false)' : 'false'};
   let isConnected = true;
 
   onMount(async () => {
@@ -198,7 +197,7 @@ Feel free to delete this file and the entire sentry route.
 
     <button
       type="button"
-      onclick={getSentryData}
+      on${options.isUsingSvelte5 ? '' : ':'}click={getSentryData}
       disabled={!isConnected}
     >
       <span>